refactor(charts.ccu): extract backend constructs
